requirements.txt
aiogram==2.25.1
deep-translator==1.11.4
SpeechRecognition==3.10.1
pydub==0.25.1
gTTS==2.5.1
SpeechRecognition==3.10.1
python-dotenv==1.0.1


